requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.0
Django==4.2.1
django-cleanup==7.0.0
django-rest-knox==4.2.0
django-unused-media==0.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-query-filter==0.1.8.5
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-magic==0.4.27
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2

coreapi~=2.3.3
coreschema~=0.0.4